Tiny TempCache + App Entry Refactor (#1)
There's still plenty this could be moved around and cleaned up and extracted to helpers but I would like to get cracking on defining new schemas so I just wanted to do this little bit.

Changes
Move github_graphql class to be named GithubGraphqlApp and treated as the entry path to running the program. Don't save state in this class at all and possibly in the future move it to a module
Temp Cache has the file age stuff moved to a constant and spread across several helpers
Add a Todo to remember what I did between stuff! I could just use github but I work without internet surprisingly often and I want to still be able to see it
Cache is not stored in git anymore (change on master already) this removes it from here